requirements.txt
openai>=1.51,<3
openai-agents==0.4.2
pydantic>=2.7,<3
pandas>=2.2,<3
scikit-learn>=1.4,<2
python-dotenv>=1.0,<2
pydantic-settings
openpyxl
fastapi>=0.115,<1
uvicorn>=0.30,<1
python-multipart>=0.0.9,<1
pyyaml
gitpython>=3.1.40
requests>=2.28.0

# CLI + logging UX
typer>=0.12.3
rich>=13.7.1

# Embedding utilities
numpy>=1.26.0